feat: Add Resolution qualifier to override BlueprintResolver
