+++ composify/attributes.py
@@ -1,15 +1,3 @@
-from dataclasses import dataclass
+from ._attributes import Name, ProvidedBy
 
-from composify.metadata.attributes import BaseAttributeMetadata
-
-
-@dataclass(frozen=True, slots=True)
-class ProvidedBy(BaseAttributeMetadata):
-    """Describe the provider of the value."""
-
-    source: str
-
-
-class Name(str, BaseAttributeMetadata):
-    def __repr__(self) -> str:
-        return f"Name({self})"
+__all__ = ("Name", "ProvidedBy")
