+++ composify/qualifiers.py
@@ -1,11 +1,13 @@
-from dataclasses import dataclass
+from composify._qualifiers import (
+    DisallowSubclass,
+    Exhaustive,
+    Resolution,
+    SelectFirst,
+)
 
-from composify.metadata.qualifiers import BaseQualifierMetadata
-
-
-@dataclass(frozen=True, slots=True)
-class DisallowSubclass(BaseQualifierMetadata):
-    disallow: bool = True
-
-    def __bool__(self) -> bool:
-        return self.disallow
+__all__ = (
+    "DisallowSubclass",
+    "Resolution",
+    "Exhaustive",
+    "SelectFirst",
+)
