+++ composify/registry.py
@@ -5,10 +5,10 @@ from collections.abc import Iterable, Sequence
 from dataclasses import dataclass
 from typing import Generic, TypeAlias, TypeVar
 
+from composify._qualifiers import DisallowSubclass
 from composify.errors import DuplicatedEntryError
 from composify.metadata.attributes import AttributeSet, collect_attributes
 from composify.metadata.qualifiers import QualifierSet, collect_qualifiers
-from composify.qualifiers import DisallowSubclass
 from composify.types import AnnotatedType, get_type, resolve_base_types
 
 Key: TypeAlias = AnnotatedType
